Move setSearchAttributes util
