+++ src/lib/services/standalone-activities.ts
@@ -22,15 +22,16 @@ import type {
 } from '$lib/types/activity-execution';
 import { activityOptionsUpdateMask } from '$lib/utilities/activity-options-update-mask';
 import { decodePayloadAndParseDataToJSON } from '$lib/utilities/decode-payload';
-import { encodePayloads } from '$lib/utilities/encode-payload';
+import {
+  encodePayloads,
+  setSearchAttributes,
+} from '$lib/utilities/encode-payload';
 import { stringifyWithBigInt } from '$lib/utilities/parse-with-big-int';
 import {
   type ErrorCallback,
   requestFromAPI,
 } from '$lib/utilities/request-from-api';
 import { routeForApi } from '$lib/utilities/route-for-api';
-
-import { setSearchAttributes } from './workflow-service';
 
 // Timeout duration inputs on the activity forms; largest-first so
 // getFirstWholeNumberUnit resolves to the coarsest whole unit, defaulting to
